requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
codewars-test-teey==0.1.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
mysql-connector==2.2.9
numpy==1.25.1
outcome==1.2.0
packaging==23.1
pandas==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
